refactor: refactor signin page
Add helpers, hooks, types to refactor signin form
