+++ src/modules/signin/components/SigninForm.helpers.ts
@@ -1,0 +1,6 @@
+import { TSigninFormFields } from "./SigninForm.types";
+
+export const signinFormInitialValues: TSigninFormFields = {
+  email: "",
+  password: "",
+};
