+++ src/modules/signin/components/SigninForm.types.ts
@@ -1,0 +1,4 @@
+export type TSigninFormFields = {
+  email: string;
+  password: string;
+};
